+++ Company-Project/src/main/pages/home.py
@@ -1,4 +1,5 @@
 from django.utils.translation import gettext_lazy as _
+from wagtail.models import PageManager
 from wagtail_headless_preview.models import HeadlessPreviewMixin
 
 from .base import BasePage
@@ -7,6 +8,8 @@ from .base import BasePage
 class HomePage(HeadlessPreviewMixin, BasePage):
     extra_panels = BasePage.extra_panels
     serializer_class = "main.pages.HomePageSerializer"
+
+    objects: PageManager
 
     class Meta:
         verbose_name = _("Home")