+++ {{cookiecutter.project_name}}/src/main/pages/article.py
@@ -1,6 +1,7 @@
 from django.utils.translation import gettext_lazy as _
-from wagtail.fields import RichTextField
 from wagtail.admin.panels import FieldPanel
+from wagtail.fields import RichTextField
+from wagtail.models import PageManager
 from wagtail_headless_preview.models import HeadlessPreviewMixin
 
 from .base import BasePage
@@ -13,6 +14,8 @@ class ArticlePage(HeadlessPreviewMixin, BasePage):
 
     extra_panels = BasePage.extra_panels
     serializer_class = "main.pages.ArticlePageSerializer"
+
+    objects: PageManager
 
     class Meta:
         verbose_name = _("Article")